Tier1 context menu ImageResizer Minor cleanups Move logic to package.h
