+++ src/modules/imageresizer/ImageResizerContextMenu/Source.def
@@ -1,0 +1,5 @@
+LIBRARY
+EXPORTS             
+DllCanUnloadNow           PRIVATE
+DllGetClassObject         PRIVATE
+DllGetActivationFactory   PRIVATE
