+++ src/modules/imageresizer/ImageResizerContextMenu/framework.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#define WIN32_LEAN_AND_MEAN             // Exclude rarely-used stuff from Windows headers
+// Windows Header Files
+#include <windows.h>
