+++ src/modules/imageresizer/ImageResizerLib/ImageResizerConstants.h
