+++ src/modules/powerrename/dll/dllmain.cpp
@@ -1,128 +1,26 @@
 #include "pch.h"
+
+#include "Generated Files/resource.h"
+#include "PowerRenameConstants.h"
 #include "PowerRenameExt.h"
+
 #include <interface/powertoy_module_interface.h>
 #include <settings.h>
 #include <trace.h>
+#include <VersionHelpers.h>
+
 #include <common/SettingsAPI/settings_objects.h>
 #include <common/logger/logger.h>
 #include <common/utils/logger_helper.h>
+#include <common/utils/package.h>
 #include <common/utils/process_path.h>
 #include <common/utils/resources.h>
-#include "Generated Files/resource.h"
+
 #include <atomic>
-#include <VersionHelpers.h>
-#include <dll/PowerRenameConstants.h>
-#include <winrt/Windows.ApplicationModel.h>
-#include <winrt/Windows.Foundation.h>
-#include <winrt/Windows.Management.Deployment.h>
 
 std::atomic<DWORD> g_dwModuleRefCount = 0;
 HINSTANCE g_hInst = 0;
 
-namespace
-{
-    BOOL IsWin11OrGreater()
-    {
-        OSVERSIONINFOEX osvi;
-        DWORDLONG dwlConditionMask = 0;
-        int op = VER_GREATER_EQUAL;
-
-        // Initialize the OSVERSIONINFOEX structure.
-
-        ZeroMemory(&osvi, sizeof(OSVERSIONINFOEX));
-        osvi.dwOSVersionInfoSize = sizeof(OSVERSIONINFOEX);
-        osvi.dwMajorVersion = HIBYTE(_WIN32_WINNT_WINTHRESHOLD);
-        osvi.dwMinorVersion = LOBYTE(_WIN32_WINNT_WINTHRESHOLD);
-        // Windows 11 build number
-        osvi.dwBuildNumber = 22000;
-
-        // Initialize the condition mask.
-        VER_SET_CONDITION(dwlConditionMask, VER_MAJORVERSION, op);
-        VER_SET_CONDITION(dwlConditionMask, VER_MINORVERSION, op);
-        VER_SET_CONDITION(dwlConditionMask, VER_BUILDNUMBER, op);
-
-        // Perform the test.
-        return VerifyVersionInfo(
-            &osvi,
-            VER_MAJORVERSION | VER_MINORVERSION | VER_BUILDNUMBER,
-            dwlConditionMask);
-    }
-
-    bool IsPackageRegistered(std::wstring packageDisplayName)
-    {
-        using namespace winrt::Windows::Foundation;
-        using namespace winrt::Windows::Management::Deployment;
-
-        PackageManager packageManager;
-
-        for (auto const& package : packageManager.FindPackagesForUser({}))
-        {
-            const auto& packageFullName = std::wstring{ package.Id().FullName() };
-
-            if (packageFullName.contains(packageDisplayName))
-            {
-                return true;
-            }
-        }
-
-        return false;
-    }
-
-    bool RegisterSparsePackage(std::wstring externalLocation, std::wstring sparsePkgPath)
-    {
-        using namespace winrt::Windows::Foundation;
-        using namespace winrt::Windows::Management::Deployment;
-
-        try
-        {
-            Uri externalUri{ externalLocation };
-            Uri packageUri{ sparsePkgPath };
-
-            PackageManager packageManager;
-
-            // Declare use of an external location
-            AddPackageOptions options;
-            options.ExternalLocationUri(externalUri);
-
-            IAsyncOperationWithProgress<DeploymentResult, DeploymentProgress> deploymentOperation = packageManager.AddPackageByUriAsync(packageUri, options);
-            deploymentOperation.get();
-
-            // Check the status of the operation
-            if (deploymentOperation.Status() == AsyncStatus::Error)
-            {
-                auto deploymentResult{ deploymentOperation.GetResults() };
-                auto errorCode = deploymentOperation.ErrorCode();
-                auto errorText = deploymentResult.ErrorText();
-
-                Logger::error(L"Register PowerRenameContextMenu package failed. ErrorCode: {}, ErrorText: {}", std::to_wstring(errorCode), errorText);
-                return false;
-            }
-            else if (deploymentOperation.Status() == AsyncStatus::Canceled)
-            {
-                Logger::error(L"Register PowerRenameContextMenu package canceled.");
-                return false;
-            }
-            else if (deploymentOperation.Status() == AsyncStatus::Completed)
-            {
-                Logger::info(L"Register PowerRenameContextMenu package completed.");
-            }
-            else
-            {
-                Logger::debug(L"Register PowerRenameContextMenu package started.");
-            }
-
-            return true;
-        }
-        catch (std::exception e)
-        {
-            std::string errorMessage{ "Exception thrown while trying to register PowerRenameContextMenu package: " };
-            errorMessage += e.what();
-            Logger::error(errorMessage);
-
-            return false;
-        }
-    }
-}
 class CPowerRenameClassFactory : public IClassFactory
 {
 public:
@@ -290,15 +188,15 @@ public:
         Logger::info(L"PowerRename enabled");
         m_enabled = true;
 
-        if (IsWin11OrGreater())
+        if (package::IsWin11OrGreater())
         {
             std::wstring path = get_module_folderpath(g_hInst);
             std::wstring packageUri = path + L"\\PowerRenameContextMenuPackage.msix";
 
             std::wstring packageDisplayName{ L"PowerRenameContextMenu" };
-            if (!IsPackageRegistered(packageDisplayName))
+            if (!package::IsPackageRegistered(packageDisplayName))
             {
-                RegisterSparsePackage(path, packageUri);
+                package::RegisterSparsePackage(path, packageUri);
             }
         }
